+++ rmake.py
@@ -2,254 +2,234 @@
 """ Copyright (c) 2021-2025 Advanced Micro Devices, Inc. All rights reserved.
 Manage build and installation"""
 
-import re
-import sys
 import os
 import subprocess
 import argparse
-import ctypes
-import pathlib
-from fnmatch import fnmatchcase
+import platform as pf
+import shutil
 
-args = {}
-param = {}
-OS_info = {}
+class AutoBuilder:
+    def __parse_args__(self):
+        """Parse command-line arguments"""
+        self.default_gpus = 'gfx906:xnack-,gfx1030,gfx1100,gfx1101,gfx1102,gfx1151,gfx1200,gfx1201'
 
-def parse_args():
-    """Parse command-line arguments"""
-    parser = argparse.ArgumentParser(description="""
-    Checks build arguments
-    """)
+        self.parser.add_argument('-a', '--architecture', dest='gpu_architecture', required=False, default=self.default_gpus, 
+                            help='Set GPU architectures, e.g. all, gfx000, gfx803, gfx906:xnack-;gfx1030;gfx1100 (optional, default: all)')
+        self.parser.add_argument('-b', '--benchmarks', required=False, default=False, dest='build_bench', action='store_true',
+                                 help='Generate benchmarks only (default: False)')
+        self.parser.add_argument('-c', '--clients', required=False, default=False, dest='build_clients', action='store_true',
+                            help='Generate all client builds (default: False)')
+        self.parser.add_argument('-g', '--debug', required=False, default=False,  action='store_true',
+                            help='Generate Debug build (default: False)')
+        self.parser.add_argument('-i', '--install', required=False, default=False, dest='install', action='store_true',
+                            help='Install after build (default: False)')
+        self.parser.add_argument('-t', '--tests', required=False, default=False, dest='build_tests', action='store_true',
+                            help='Generate unit tests only (default: False)')
+        self.parser.add_argument('-v', '--verbose', required=False, default=False, action='store_true',
+                            help='Verbose build (default: False)')  
+        self.parser.add_argument(      '--build_dir', type=str, required=False, default="build",
+                            help='Build directory path (default: build)')
+        self.parser.add_argument(      '--cmake-darg', required=False, dest='cmake_dargs', action='append', default=[],
+                            help='List of additional cmake defines for builds (e.g. CMAKE_CXX_COMPILER_LAUNCHER=ccache)')
+        self.parser.add_argument(      '--deps_dir', type=str, required=False, default=None,
+                            help='Dependencies directory path (default: build/deps)')
+        self.parser.add_argument(      '--skip_ld_conf_entry', required=False, default=False)
+        self.parser.add_argument(      '--static', required=False, default=False, dest='static_lib', action='store_true',
+                            help='Generate static library build (default: False)')
+        
+        self.args = self.parser.parse_args()
 
-    default_gpus = 'gfx906:xnack-,gfx1030,gfx1100,gfx1101,gfx1102,gfx1151,gfx1200,gfx1201'
+    def __init__(self):
+        self.parser = argparse.ArgumentParser(description="Checks build arguments")
+        self.__parse_args__()
+        
+        sysInfo = pf.uname()
 
-    parser.add_argument('-g', '--debug', required=False, default=False,  action='store_true',
-                        help='Generate Debug build (default: False)')
-    parser.add_argument(      '--build_dir', type=str, required=False, default="build",
-                        help='Build directory path (default: build)')
-    parser.add_argument(      '--deps_dir', type=str, required=False, default=None,
-                        help='Dependencies directory path (default: build/deps)')
-    parser.add_argument(      '--skip_ld_conf_entry', required=False, default=False)
-    parser.add_argument(      '--static', required=False, default=False, dest='static_lib', action='store_true',
-                        help='Generate static library build (default: False)')
-    parser.add_argument('-c', '--clients', required=False, default=False, dest='build_clients', action='store_true',
-                        help='Generate all client builds (default: False)')
-    parser.add_argument('-t', '--tests', required=False, default=False, dest='build_tests', action='store_true',
-                        help='Generate unit tests only (default: False)')
-    parser.add_argument('-i', '--install', required=False, default=False, dest='install', action='store_true',
-                        help='Install after build (default: False)')
-    parser.add_argument(      '--cmake-darg', required=False, dest='cmake_dargs', action='append', default=[],
-                        help='List of additional cmake defines for builds (e.g. CMAKE_CXX_COMPILER_LAUNCHER=ccache)')
-    parser.add_argument('-a', '--architecture', dest='gpu_architecture', required=False, default=default_gpus, #:sramecc+:xnack-" ) #gfx1030" ) #gfx906" ) # gfx1030" )
-                        help='Set GPU architectures, e.g. all, gfx000, gfx803, gfx906:xnack-;gfx1030;gfx1100 (optional, default: all)')
-    parser.add_argument('-v', '--verbose', required=False, default=False, action='store_true',
-                        help='Verbose build (default: False)')
-    return parser.parse_args()
+        # getting os information
+        self.OS_info = {
+            "Machine" : sysInfo.machine,
+            "Node Name" : sysInfo.node,
+            "Num Processor" : os.cpu_count(),
+            "Processor" : sysInfo.processor,
+            "Release" : sysInfo.release,
+            "System" : sysInfo.system,
+            "Version" : sysInfo.version,
+        }
 
-def os_detect():
-    inf_file = "/etc/os-release"
-    if os.path.exists(inf_file):
-        with open(inf_file) as f:
-            for line in f:
-                if "=" in line:
-                    k,v = line.strip().split("=")
-                    OS_info[k] = v.replace('"','')
-    else:
-        OS_info["ID"] = 'windows'
-        OS_info["VERSION_ID"] = 10
-    OS_info["NUM_PROC"] = os.cpu_count()
-    print(OS_info)
+        self.custom_cmake_args = set()
+        self.cmake_options = []
 
-def create_dir(dir_path):
-    if os.path.isabs(dir_path):
-        full_path = dir_path
-    else:
-        fullpath = os.path.join( os.getcwd(), dir_path )
-    pathlib.Path(fullpath).mkdir(parents=True, exist_ok=True)
-    return
+        for item in self.args.cmake_dargs:
+            self.custom_cmake_args.add(item.split('=')[0]) # get the argument name
 
-def delete_dir(dir_path) :
-    if (not os.path.exists(dir_path)):
-        return
-    if (OS_info["ID"] == 'windows'):
-        run_cmd( "RMDIR" , f"/S /Q {dir_path}")
-    else:
-        linux_path = pathlib.Path(dir_path).absolute()
-        run_cmd( "rm" , f"-rf {linux_path}")
+        self.lib_dir = os.path.dirname(os.path.abspath(__file__)) 
+        self.toolchain = f'toolchain-linux.cmake' if self.OS_info['System'] == 'Linux' else f'toolchain-windows.cmake'
 
-def cmake_path(os_path):
-    if OS_info["ID"] == "windows":
-        return os_path.replace("\\", "/")
-    else:
-        return os.path.realpath(os_path)
-
-def config_cmd():
-    global args
-    global OS_info
-    cwd_path = os.getcwd()
-    src_path = cwd_path.replace("\\", "/")
-
-    print( f"***************** {src_path}")
-    cmake_executable = ""
-    cmake_options = []
-    cmake_platform_opts = []
-    if (OS_info["ID"] == 'windows'):
-        # we don't have ROCM on windows but have hip, ROCM can be downloaded if required
-        # CMAKE_PREFIX_PATH set to rocm_path and HIP_PATH set BY SDK Installer
-        raw_rocm_path = cmake_path(os.getenv('HIP_PATH', "C:/hip"))
-        rocm_path = f'"{raw_rocm_path}"' # guard against spaces in path
-        cmake_executable = "cmake.exe"
-        toolchain = os.path.join( src_path, "toolchain-windows.cmake" )
-        #set CPACK_PACKAGING_INSTALL_PREFIX= defined as blank as it is appended to end of path for archive creation
-        cmake_platform_opts.append( f"-DWIN32=ON -DCPACK_PACKAGING_INSTALL_PREFIX=") #" -DCPACK_PACKAGING_INSTALL_PREFIX={rocm_path}"
-        cmake_platform_opts.append( f"-DCMAKE_INSTALL_PREFIX=\"C:/hipSDK\"" )
-
-        # MSVC requires acknowledgement of using extended aligned storage.
-        # Before VS 2017 15.8, has non-conforming alignment. VS 2017 15.8 fixes this, but inherently changes layouts of
-        # aligned storage with extended alignment, and thus binary compatibility with such types.
-        cmake_platform_opts.append( "-DCMAKE_CXX_FLAGS=\"-D_ENABLE_EXTENDED_ALIGNED_STORAGE\"")
-
-        rocm_cmake_path = '"' + cmake_path(os.getenv("ROCM_CMAKE_PATH", "C:/hipSDK")) + '"'
-        generator = f"-G Ninja"
-        # "-G \"Visual Studio 16 2019\" -A x64"  #  -G NMake ")  #
-        cmake_options.append( generator )
-    else:
-        rocm_path = os.getenv( 'ROCM_PATH', "/opt/rocm")
-        rocm_cmake_path = '"' + rocm_path + '"'
-        if (OS_info["ID"] in ['centos', 'rhel']):
-          cmake_executable = "cmake3"
+    def __mk_dir__(self, dir_path: str):
+        if os.path.isabs(dir_path):
+            full_path = dir_path
         else:
-          cmake_executable = "cmake"
-        toolchain = "toolchain-linux.cmake"
-        cmake_platform_opts = [f"-DROCM_DIR:PATH={rocm_path}", f"-DCPACK_PACKAGING_INSTALL_PREFIX={rocm_path}"]
+            full_path = os.path.join(self.lib_dir, dir_path)
 
-    tools = f"-DCMAKE_TOOLCHAIN_FILE={toolchain}"
-    cmake_options.append( tools )
+        try:
+            os.makedirs(full_path)
+        except FileExistsError:
+            ... # file already exists
 
-    cmake_options.extend( cmake_platform_opts)
+    def __rm_dir__(self, dir_path: str):
+        if os.path.isabs(dir_path):
+            full_path = dir_path
+        else:
+            full_path = os.path.join(self.lib_dir, dir_path)
+        try:
+            if self.OS_info['System'] == 'Linux':
+                subprocess.run(f'rm -rf "{full_path}"', shell=True)
+            else:
+                subprocess.run(f'RMDIR /S /Q {full_path}', shell=True)
 
+        except FileNotFoundError:
+            ... # no file to remove
 
-  # build type
-    cmake_config = ""
-    build_dir = args.build_dir
-    if not args.debug:
-        build_path = os.path.join(build_dir, "release")
-        cmake_config="Release"
-    else:
-        build_path = os.path.join(build_dir, "debug")
-        cmake_config="Debug"
-
-    cmake_options.append( f"-DCMAKE_BUILD_TYPE={cmake_config}" ) #--build {build_path}" )
-
-    if args.deps_dir is None:
-        deps_dir = os.path.abspath(os.path.join(build_dir, 'deps')).replace('\\','/')
-    else:
-        deps_dir = args.deps_dir
-
-    if (OS_info["ID"] == 'windows'):
-        cmake_base_options = f"-DROCM_PATH={rocm_path} -DCMAKE_PREFIX_PATH:PATH={rocm_path[:-1]};{rocm_cmake_path[1:]}" # -DCMAKE_INSTALL_PREFIX=rocmath-install" #-DCMAKE_INSTALL_LIBDIR=
-    else:
-        cmake_base_options = f"-DROCM_PATH={rocm_path} -DCMAKE_PREFIX_PATH:PATH={rocm_path[:-1]},{rocm_cmake_path[1:-1]}" # -DCMAKE_INSTALL_PREFIX=rocmath-install" #-DCMAKE_INSTALL_LIBDIR=
+    def __insert_cmake_args__(self, arg, val):
+        # This is to prevent the case of having duplicate cmake args from the argument --cmake-darg
+        if arg not in self.custom_cmake_args:
+            self.cmake_options.append(f'-D{arg}={val}')
     
-    cmake_options.append( cmake_base_options )
+    def __get_cmake_cmd__(self):
 
-    print( cmake_options )
+        m = ' Current Working Directory '
+        print(f'{m:-^100}\n\t{self.lib_dir}')
+        print()
 
-    # clean
-    delete_dir( build_path )
+        self.__insert_cmake_args__('CMAKE_TOOLCHAIN_FILE', self.toolchain)
+        self.__insert_cmake_args__('CPACK_INCLUDE_TOPLEVEL_DIRECTORY', 'OFF')
+        self.__insert_cmake_args__('CPACK_SET_DESTDIR', 'OFF')
+        self.__insert_cmake_args__('GPU_TARGETS', self.args.gpu_architecture)
+ 
+        cmake_exe = ''
+        if self.args.debug:
+            build_path = os.path.join(self.args.build_dir, 'debug')
+            cmake_config = 'Debug'
+        else:
+            build_path = os.path.join(self.args.build_dir, 'release')
+            cmake_config = 'Release'
+        
+        self.build_path = build_path
+        
+        self.cmake_options.append(f"-DCMAKE_BUILD_TYPE={cmake_config}")
 
-    create_dir( os.path.join(build_path, "clients") )
-    os.chdir( build_path )
+        if self.OS_info['System'] == 'Linux':
+            cmake_exe = shutil.which('cmake3')
 
-    # packaging options
-    cmake_pack_options = f"-DCPACK_SET_DESTDIR=OFF -DCPACK_INCLUDE_TOPLEVEL_DIRECTORY=OFF"
-    cmake_options.append( cmake_pack_options )
+            if cmake_exe is None:
+                cmake_exe = shutil.which('cmake')
+            
+            if cmake_exe is None:
+                raise(SystemError('Did not find cmake or cmake3 in system'))
 
-    if args.static_lib:
-        cmake_options.append( f"-DBUILD_SHARED_LIBS=OFF" )
+            rocm_path = os.getenv('ROCM_PATH', '/opt/rocm')
 
-    if args.skip_ld_conf_entry:
-        cmake_options.append( f"-DROCM_DISABLE_LDCONFIG=ON" )
+            self.__insert_cmake_args__('CMAKE_CXX_FLAGS', f'"-w"')
+            self.__insert_cmake_args__('CMAKE_PREFIX_PATH:PATH', rocm_path)
+            self.__insert_cmake_args__('CPACK_PACKAGING_INSTALL_PREFIX', rocm_path)
+            self.__insert_cmake_args__('ROCM_DIR:PATH', rocm_path)
+        else:
+            cmake_exe = shutil.which('cmake.exe')
+            
+            if cmake_exe is None:
+                cmake_exe = shutil.which('cmake3.exe')
 
-    if args.build_tests:
-        cmake_options.append( f"-DBUILD_TEST=ON -DBUILD_DIR={build_dir}" )
+            if cmake_exe is None:
+                raise(SystemError('Did not find cmake or cmake3 in system'))
+            
+            rocm_path = os.getenv('ROCM_PATH', 'C:/hip')
+            rocm_cmake_path = os.getenv('ROCM_CMAKE_PATH', r'C:/hipSDK')
 
-    if args.build_clients:
-        cmake_options.append( f"-DBUILD_TEST=ON -DBUILD_BENCHMARK=ON -DBUILD_EXAMPLE=ON -DBUILD_DIR={build_dir}" )
+            rocm_path.replace('\\', '/')
+            rocm_cmake_path.replace('\\', '/')
 
-    cmake_options.append( f"-DAMDGPU_TARGETS={args.gpu_architecture}" )
+            if '-G Ninja' not in self.custom_cmake_args:
+                self.cmake_options.append(f'-G Ninja')
+            self.__insert_cmake_args__('WIN32', 'ON')
+            self.__insert_cmake_args__('CPACK_PACKAGING_INSTALL_PREFIX', '')            
+            self.__insert_cmake_args__('CMAKE_INSTALL_PREFIX', rocm_cmake_path)            
+            self.__insert_cmake_args__('CMAKE_CXX_FLAGS', f'"-D_ENABLE_EXTENDED_ALIGNED_STORAGE -w"')
+            self.__insert_cmake_args__('CMAKE_PREFIX_PATH:PATH', f'{rocm_path};{rocm_cmake_path}')
 
-    if args.cmake_dargs:
-        for i in args.cmake_dargs:
-          cmake_options.append( f"-D{i}" )
+        self.__insert_cmake_args__('ROCM_PATH', rocm_path)
 
-    cmake_options.append( f"{src_path}")
+        if self.args.static_lib: self.__insert_cmake_args__('BUILD_SHARED_LIBS', 'OFF')
+        if self.args.skip_ld_conf_entry: self.__insert_cmake_args__('ROCM_DISABLE_LDCONFIG', 'ON')
+        
+        if self.args.build_clients:
+            self.args.build_tests = True
+            self.args.build_bench = True
+            self.__insert_cmake_args__('BUILD_EXAMPLE', 'ON')
 
-#   case "${ID}" in
-#     centos|rhel)
-#     cmake_options="${cmake_options} -DCMAKE_FIND_ROOT_PATH=/usr/lib64/llvm7.0/lib/cmake/"
-#     ;;
-#     windows)
-#     cmake_options="${cmake_options} -DWIN32=ON -DROCM_PATH=${rocm_path} -DROCM_DIR:PATH=${rocm_path} -DCMAKE_PREFIX_PATH:PATH=${rocm_path}"
-#     cmake_options="${cmake_options} --debug-trycompile -DCMAKE_MAKE_PROGRAM=nmake.exe -DCMAKE_TOOLCHAIN_FILE=toolchain-windows.cmake"
-#     # -G '"NMake Makefiles JOM"'"
-#     ;;
-#   esac
-    cmd_opts = " ".join(cmake_options)
+        if self.args.build_tests: self.__insert_cmake_args__('BUILD_TEST', 'ON')
+        if self.args.build_bench: self.__insert_cmake_args__('BUILD_BENCHMARK', 'ON')
 
-    return cmake_executable, cmd_opts
+        if self.args.cmake_dargs:
+            self.cmake_options += [f'-D{i}' for i in self.args.cmake_dargs]
+        
+        # putting '' around paths to avoid white space in pathing
+        if self.OS_info['System'] == 'Linux':
+            command_str = f"{cmake_exe}"  
+        else:
+            command_str = f'"{cmake_exe}"'
 
+        m = ' CMake Options '
+        print(f'{m:-^100}')
+        for op in self.cmake_options:
+            print(f'\t{op}')
+            command_str += f' {op}'
+        print()
 
-def make_cmd():
-    global args
-    global OS_info
+        command_str += f' "{self.lib_dir}"'
+        m = ' Final Command '
+        print(f'{m:-^100}')
+        print(command_str)
+        print()
+        return command_str
+    
+    def __call__(self):
+        m = ' System Information '
+        print()
+        print(f'{m:-^100}')
+        
+        for k in self.OS_info:
+            print(f'\t {k}: {self.OS_info[k]}')
+        print()
 
-    make_options = []
+        cmake_command = self.__get_cmake_cmd__()
 
-    if (OS_info["ID"] == 'windows'):
-        make_executable = "cmake.exe --build ." # ninja"
-        if args.verbose:
-          make_options.append( "--verbose" )
-        make_options.append( "--target all" )
-        if args.install:
-          make_options.append( "--target package --target install" )
-    else:
-        nproc = OS_info["NUM_PROC"]
-        make_executable = f"make -j {nproc}"
-        if args.verbose:
-          make_options.append( "VERBOSE=1" )
-        if args.install:
-          make_options.append( "install" )
-    cmd_opts = " ".join(make_options)
+        self.__rm_dir__(self.build_path)
+        self.__rm_dir__('build')
+        self.__mk_dir__(self.build_path)
 
-    return make_executable, cmd_opts
+        curr_dir = os.path.abspath(os.curdir)
+        os.chdir(self.build_path)
 
-def run_cmd(exe, opts):
-    program = f"{exe} {opts}"
-    if sys.platform.startswith('win'):
-        sh = True
-    else:
-        sh = True
-    print(program)
-    proc = subprocess.run(program, check=True, stderr=subprocess.STDOUT, shell=sh)
-    #proc = subprocess.Popen(cmd, cwd=os.getcwd())
-    #cwd=os.path.join(workingdir,"..",".."), stdout=fout, stderr=fout,
-     #                       env=os.environ.copy())
-    #proc.wait()
-    return proc.returncode
+        subprocess.run(cmake_command, shell=True)
+        
+        if self.OS_info['System'] == 'Linux':
+            v = ''
+            if self.args.verbose:
+                v = ' VERBOSE=1'
+            subprocess.run(f' make -j {self.OS_info["Num Processor"]}{v}', shell=True)
 
-def main():
-    global args
-    os_detect()
-    args = parse_args()
-    # configure
-    exe, opts = config_cmd()
-    run_cmd(exe, opts)
+            if self.args.install:
+                subprocess.run(f'make install', shell=True)
+        else:
+            v = ''
+            if self.args.verbose:
+                v = ' --verbose'
+            subprocess.run(f'ninja -j {self.OS_info["Num Processor"]}{v}', shell=True)
+            if self.args.install:
+                subprocess.run(f'ninja install', shell=True)    
 
-    # make/build/install
-    exe, opts = make_cmd()
-    run_cmd(exe, opts)
-
+        os.chdir(curr_dir)
 
 if __name__ == '__main__':
-    main()
+    builder = AutoBuilder()
+    builder()